Extract TaskList from Bob.java
